Fix primitive class handling in Serialisers
Fixes: ##37776
