+++ independent-projects/resteasy-reactive/common/runtime/src/main/java/org/jboss/resteasy/reactive/common/core/Serialisers.java
@@ -120,7 +120,7 @@ public abstract class Serialisers {
         if (Response.class.isAssignableFrom(entityType)) {
             return Collections.emptyList();
         }
-        Class<?> klass = primitiveWrapperOf(entityType);
+        Class<?> klass = lookupPrimitiveWrapper(entityType);
         //first we check to make sure that the return type is build time selectable
         //this fails when there are eligible writers for a sub type of the entity type
         //e.g. if the entity type is Object and there are mappers for String then we
@@ -235,7 +235,28 @@ public abstract class Serialisers {
         return findWriters(configuration, entityType, resolvedMediaType, null);
     }
 
-    protected Class<?> lookupPrimitiveWrapper(Class<?> entityType) {
+    protected final Class<?> lookupPrimitiveWrapper(final Class<?> entityType) {
+        if (!entityType.isPrimitive()) {
+            return entityType;
+        }
+        if (entityType == boolean.class) {
+            return Boolean.class;
+        } else if (entityType == char.class) {
+            return Character.class;
+        } else if (entityType == byte.class) {
+            return Byte.class;
+        } else if (entityType == short.class) {
+            return Short.class;
+        } else if (entityType == int.class) {
+            return Integer.class;
+        } else if (entityType == long.class) {
+            return Long.class;
+        } else if (entityType == float.class) {
+            return Float.class;
+        } else if (entityType == double.class) {
+            return Double.class;
+        }
+        // this shouldn't really happen, but better be safe than sorry
         return entityType;
     }
 
@@ -244,8 +265,7 @@ public abstract class Serialisers {
         // FIXME: invocation is very different between client and server, where the server doesn't treat GenericEntity specially
         // it's probably missing from there, while the client handles it upstack
         List<MediaType> mt = Collections.singletonList(resolvedMediaType);
-        Class<?> klass = entityType;
-        klass = primitiveWrapperOf(entityType);
+        Class<?> klass = lookupPrimitiveWrapper(entityType);
         QuarkusMultivaluedMap<Class<?>, ResourceWriter> writers;
         if (configuration != null && !configuration.getResourceWriters().isEmpty()) {
             writers = new QuarkusMultivaluedHashMap<>();
@@ -258,10 +278,6 @@ public abstract class Serialisers {
         }
 
         return toMessageBodyWriters(findResourceWriters(writers, klass, mt, runtimeType));
-    }
-
-    private Class<?> primitiveWrapperOf(Class<?> entityType) {
-        return entityType;
     }
 
     public static class Builtin {
