+++ independent-projects/resteasy-reactive/server/runtime/src/main/java/org/jboss/resteasy/reactive/server/core/ServerSerialisers.java
@@ -178,32 +178,6 @@ public class ServerSerialisers extends Serialisers {
         }
     };
 
-    @Override
-    protected final Class<?> lookupPrimitiveWrapper(final Class<?> entityType) {
-        if (!entityType.isPrimitive()) {
-            return entityType;
-        }
-        if (entityType == boolean.class) {
-            return Boolean.class;
-        } else if (entityType == char.class) {
-            return Character.class;
-        } else if (entityType == byte.class) {
-            return Byte.class;
-        } else if (entityType == short.class) {
-            return Short.class;
-        } else if (entityType == int.class) {
-            return Integer.class;
-        } else if (entityType == long.class) {
-            return Long.class;
-        } else if (entityType == float.class) {
-            return Float.class;
-        } else if (entityType == double.class) {
-            return Double.class;
-        }
-        // this shouldn't really happen, but better be safe than sorry
-        return entityType;
-    }
-
     public static boolean invokeWriter(ResteasyReactiveRequestContext context, Object entity, MessageBodyWriter writer,
             ServerSerialisers serialisers)
             throws IOException {
